refactor: update index.ts to remove requestHandler function to middleware
